requirements.txt
transformers==4.42.4
datasets==2.20.0
accelerate==0.32.1
bitsandbytes==0.43.1
peft==0.12.0
sentencepiece==0.2.0
protobuf==4.25.3
pillow==10.4.0 # for loading images with the datasets library
nltk==3.8.1
lightning==2.3.3
tensorboard==2.17.0
torch_tb_profiler==0.4.3
xlsxwriter==3.2.0
pytest==8.3.2
scikit-learn==1.5.1